Fix doc problems
Math in docstrings needs to be done in a special way to work.
Remove symbol exported from ngraph that didn't exist.
Comment out svg inclusion for now, hoping that they fix the bug. Could also use a raw html object inclusion,
but I don't know how to get the .svg file over to the build directory for inclusion that way.