Add ng.sequential and ng.sequential_factory (syntactic sugar)
ng.sequential specifies an order for executing ops, and has as a value
the last op (which may also occur earlier in the list; ops only get executed
once). This makes state side-effects deterministic.

Modify optimizers to use ng.sequential_factory. This revealed an error
in the optimizers where the learning rate isn't updated, as described
in issue #763. An optimizer method is added to update the rate, but it
is only being called from the optimizer test.