Remove .pyc before running tests, ignore test outputs
The docker image has different pathnames than the local files, and
pytest checks that they match, so remove them before running tests.
This may be a good idea in general in case file timestamps get confused
and cause a .pyc to not correspond to the .py file, making a test pass
that shouldn't.