Tests to check Gather Op is inserted
test graph with X, Y,and Z, where Z, X is on device 0, but Y is
split(parallelized) between device 1 and 2.

A Computation op that depends on Y should insert a GatherRecv op
into the graph.

We could use a mock here to test for the graph we want to
compute