Fix absolute tolerance value for ngraph unit tests, fix derivative tests
Fixed tests are:
- py.test tests/transformer_dependent/test_execution.py -s -k 'test_elementwise_binary_ops_matched_args_deriv_lhs or test_elementwise_binary_ops_matched_args_deriv_rhs' --transformer=argon

These are derivative tests that should not be calculated in flexpoint as it has too poor accuracy.
'Argon' transformer was added to flex_transformers list.

- py.test -s tests/transformer_dependent/test_kernel_generator.py --transformer=argon, except case 'test_kernel_cache'

Abolute tolerance for different test cases can differ in few orders of magnitude.
Setting one value for all of them is not good idea as it has to be the largest
that suits all the cases. Than it would be far too large for cases where small
accuracy is enough and will not detect wrong computations.

Flexpoint accuracy depends on maximum element of tensor.
Ngraph tests I looked at use values in the range [0, 5] so dec is usually around 12.
It gives absolute error of single number 0.5*2^-12 = 0.00012
Thus basic arithmetic operation require accuracy about atol=0.0005.
Reduction operation of N elements requires atol N times larger.
Then absolute error of the result is larger by the same factor and so atol needs to be.
Reciprocal operation gives larger error for small arguments as such value has large relative error
due to the fact that is absolute value is not much higher than its absolute error.

The changes:

1. Limiting maximum absolute value for tensors elements to 7.9.

There is used np.random.randn function to fill tensors with random values. It can give any value as a result
however values above 5 are highly improbable and would appear very rarely.
I haven't ever obtained value higher than 5 in my experiments.
Limit 7.9 would almost never modify the tested tensor but would prevent from random failures from
time to time when the test is run in continuous environment.
This limit is approximate upper bound of range [4, 8). Numbers from this region can be expressed by flexpoint
number of the same dec.
Why not 15.9 that is approximate limit of [8, 16) range ?
Numbers above 8 are highly improbable and if appear from time to time can cause random failures due to
reduced accuracy of all numbers in tensor. Most numbers in normal distribution are close to 0.

2. Limitting minimum absolute value for tensors being input to reciprocal operation to 1/7.9

This is consequence of the above and flexpoint accuracy.
Numbers very small have poor absolute accuracy. When reciprocal of them is calculated the results becomes very large
and has even worse accuracy. When small numbers would be accepted as an input to reciprocal in the test the
absolute maximum value of the result is undefined and so absolute tolerance.
To have possibility to set atol in the test and test could pass with it minimum element of the tensor that is
input to reciprocal operation has to be limited.

3. Introduction of new argument 'atol_multiplier' to ngraph.assert_allclose

It can realize different absolute tolerance for different operations that require it.
It is different then default 1 for:
- number of reduced elements for reduction operation
- empirical 15 for reciprocal operation