Updates to ConvBase and axis to use roles
Trying to move ConvBase to a more axis name and order independent framing. Also trying to make it more user friendly by removing `convparams` and moving toward individual dictionaries with simpler keys. Should address #2054. This requires having clear ways of marking axes as "width", "height", and "depth", so I've reintroduced the concept of roles. They are many-to-many relationships that are tied to context and not to the axis itself.