Update README.md
NamesDB tells us that the only acceptable use when "Intel Nervana" is followed by the word "Graph" is to capitalize the word Graph.  See email thread for screencap.  The other change is for consistency: we are trying to be consistent with a word-wrap line length of no more than 80 characters.  The reason here is that when somebody clones the repo (who maybe isn't reading it directly on Github)  they can just do
```$ cat README.md```

And read output without getting a messy horizontal scroll bar to read.  :)